Split off read_lin_block functions to a separate file. Renamed function for consistency.
